resolved comments business logic

+++ 02. Business/NewHabr.Business/Services/CommentService.cs
@@ -2,6 +2,7 @@
 using NewHabr.Domain.Contracts;
 using NewHabr.Domain.Contracts.Services;
 using NewHabr.Domain.Dto;
+using NewHabr.Domain.Exceptions;
 using NewHabr.Domain.Models;
 
 namespace NewHabr.Business.Services;
@@ -19,19 +20,28 @@ public class CommentService : ICommentService
     public async Task CreateAsync(CreateCommentRequest data, CancellationToken cancellationToken = default)
     {
         var newComment = _mapper.Map<Comment>(data);
+        newComment.CreatedAt = DateTimeOffset.UtcNow;
+        newComment.ModifiedAt = DateTimeOffset.UtcNow;
+
         _repositoryManager.CommentRepository.Create(newComment);
         await _repositoryManager.SaveAsync(cancellationToken);
     }
 
-    public async Task DeleteAsync(CommentDto data, CancellationToken cancellationToken = default)
+    public async Task DeleteAsync(Guid id, CancellationToken cancellationToken = default)
     {
-        var deleteComment = _mapper.Map<Comment>(data);
+        var deleteComment = await _repositoryManager.CommentRepository.GetByIdAsync(id, cancellationToken: cancellationToken);
+        if (deleteComment is null)
+        {
+            throw new CommentNotFoundException();
+        }
+
         _repositoryManager.CommentRepository.Delete(deleteComment);
         await _repositoryManager.SaveAsync(cancellationToken);
     }
 
     public async Task<IReadOnlyCollection<CommentDto>> GetAllAsync(CancellationToken cancellationToken = default)
     {
+
         var comments = await _repositoryManager.CommentRepository.GetAllAsync(cancellationToken: cancellationToken);
         return _mapper.Map<List<CommentDto>>(comments);
     }
@@ -48,9 +58,18 @@ public class CommentService : ICommentService
         return _mapper.Map<List<CommentDto>>(comments);
     }
 
-    public async Task UpdateAsync(CommentDto data, CancellationToken cancellationToken = default)
+    public async Task UpdateAsync(CommentDto commentDto, CancellationToken cancellationToken = default)
     {
-        var updateComment = _mapper.Map<Comment>(data);
+        var updateComment = await _repositoryManager.CommentRepository.GetByIdAsync(commentDto.Id, cancellationToken: cancellationToken);
+        if (updateComment is null)
+        {
+            throw new CommentNotFoundException();
+        }
+
+        _mapper.Map(commentDto, updateComment);
+
+        updateComment.ModifiedAt = DateTimeOffset.UtcNow;
+
         _repositoryManager.CommentRepository.Update(updateComment);
         await _repositoryManager.SaveAsync(cancellationToken);
     }
